Move /foos specific code to foos folder
